Allow passing in custom sampler in ReplayBuffer
Summary:
This diff allows passing in a custom sampler in the ReplayBuffer class. The changes include adding a new sampler class and modifying the ReplayBuffer class to accept a sampler as a parameter. The code changes also include modifying the test_replay_buffer.py file to test the new sampler functionality.



Test Plan:
unit tests (flaky)