Document reference Dockerfile
I don't know if the forge project wants to actually maintain a
Dockerfile, but this one worked for me as of commit
`bcd86f00f967b5c7d833059095ae3a8ba126e28a` and I wanted to at least open
the PR so that it's documented somewhere.

Maybe in the future someone else will want to run forge as a container
also.